+++ go.mod
@@ -2,10 +2,13 @@ module github.com/jstaf/onedriver
 
 require (
 	github.com/etcd-io/bbolt v1.3.3
-	github.com/hanwen/go-fuse v0.0.0-20190111173210-425e8d5301f6
+	github.com/hanwen/go-fuse v1.0.0
+	github.com/hanwen/go-fuse/v2 v2.0.2
+	github.com/kylelemons/godebug v1.1.0 // indirect
 	github.com/petermattis/goid v0.0.0-20180202154549-b0b1615b78e5 // indirect
 	github.com/sasha-s/go-deadlock v0.2.0
 	github.com/sirupsen/logrus v1.4.2
 	github.com/spf13/pflag v1.0.3
 	go.etcd.io/bbolt v1.3.3 // indirect
+	golang.org/x/sys v0.0.0-20190907184412-d223b2b6db03 // indirect
 )
